requirements.txt
moviepy>=1.0.3
Pillow>=9.0.0
numpy>=1.21.0
gTTS>=2.3.1
